Extend plotting functionalities to more transformers and other modules (#362)
* start plotting for images

* Add the image plotting  file

* Add an imagePlotMixin subclassing the ImagePlot for theimage submodule, add persistenceImage plots

* Add plot_persistence_image to the package

* Add plotting functionalities to filtrations in the image module and forbetticurves

* Preprocessing covered by plots

* Add basic plotting functions for different vectorisation  methods

* Incorporate plotting for diagram representations and images/preprocessing

* Change the docstring for ImagePlotterMixin and add fail on 3d  images

* Correct the samplingsin features, and titles

* Small wording fix

* Change dosctrings for plot in diagrams/features

* Fix typo in variable name

* Change plotting of  heatkernel to the same as persistenceimage, use  samplings

* Cast  images to floats, so as they can be always plotted

* Add a Homology and PointCloud Mixin

* Use the new PlotMixins

* Add todo in Forgetdimensions, in relation to plotting

* Add plotting to point_cloudsrescaling, correctimports

* Wording changes

* Correct n_values to n_bins and  add 's' to n_homologydimensionS

* Correct spelling mistakes

* Change the name of ...PlotterMixin to ...Plotter

* Make SlidingWindowEmbedding a PointCloudPlotter

* Fix typoin import

* Fix linting

* Correct the  way the distance matrices are displayed, and add the  SlidingWindow.plot method

* Change behaviour of transform_plot to return the 0-th sample

* Docstring fixes

* Create diagrams/representations.py, and migrate amplitude to features.py

This change reflects the new documentation structure and the difference now drawn there between feature creation for diagrams (scalars) and vector representations/vectorizations. These changes are mostly invisible to the user.

* Remove unused identity function

* Rename HomologyPlotter as DiagramPlotter

* Add homology_dimensions kwarg to DiagramPlotter, revise docstring

* Small docstring fixes

* Try adding triggers in the yaml

* Remove excessive blank line

* Revert "Change behaviour of transform_plot to return the 0-th sample"

This reverts commit 2fa0209e5ed4d3f241371618f8b3b0a864b25e83.

* Fix docstring for PlotterMixin

* Further docstring refinements in `PlotterMixin`

* Remove `DiagramPlotter`, differentiate between implementations of `plot`

Also make plot methods static in classes in `gtda.homology` and in ForgetDimensino

* Correctly import PlotterMixin

* Correct the plotting in the lorentz notebook

* Remove PointCloudPlotter and refine plot methods for point clouds

* Thorough refactoring of plotting API

- Eliminate some top-level functions for diagram representations, leave only betti_curves and betti_surfaces
- Remove ImagePlotter mixin and plotting/_plot.py, create top-level plot_heatmap function
- Rename plotting/diagrams.py to plotting/diagrams_representations.py
- Rename plotting/image.py to plotting/images.py
- Rename plotting/homology.py to plotting/persistence_digrams.py
- Fix problem with Binarizer coming from now-removed normalize parameter
- Refactor notebooks to use plotting API
- MIscellaneous small changes elsewhere

The philosophy behind the file changes in gtda/plotting is that file names now reflect data structures and do not try to follow the names of subpackages like diagrams, image, etc.

* Grammar

* Fix default parameter for homology_dimension_ix in HeatKernel

* Change plot_diagrams to allow ForgetDimension to work

* Allow _subdiagrams to better handle infs of different kind

* Add inline comment for future selves

* Write check_diagrams docstring and make it convert to float, revert previous changes to _subdiagrams

* Pass copy=False to astype in check_diagrams
