+++ gtda/homology/simplicial.py
@@ -10,19 +10,19 @@ from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.metrics.pairwise import pairwise_distances
 from sklearn.utils.validation import check_array, check_is_fitted
 
-from ..base import PlotterMixin
 from ._utils import _postprocess_diagrams
+from ..base import PlotterMixin
 from ..externals.python import ripser, SparseRipsComplex, CechComplex
+from ..plotting import plot_diagram
 from ..utils._docs import adapt_fit_transform_docs
 from ..utils.intervals import Interval
 from ..utils.validation import validate_params
-from ..plotting.homology import plot_diagram
 
 
 @adapt_fit_transform_docs
 class VietorisRipsPersistence(BaseEstimator, TransformerMixin, PlotterMixin):
     """:ref:`Persistence diagrams <persistence diagram>` resulting from
-    :ref:`Vietoris-Rips filtrations <vietoris-rips complex and vietoris-rips \
+    :ref:`Vietoris–Rips filtrations <vietoris-rips complex and vietoris-rips \
     homology>`.
 
     Given a :ref:`point cloud <finite metric spaces and point clouds>` in
@@ -60,7 +60,7 @@ class VietorisRipsPersistence(BaseEstimator, TransformerMixin, PlotterMixin):
         :math:`p` equals `coeff`.
 
     max_edge_length : float, optional, default: ``numpy.inf``
-        Upper bound on the maximum value of the Vietoris-Rips filtration
+        Upper bound on the maximum value of the Vietoris–Rips filtration
         parameter. Points whose distance is greater than this value will
         never be connected by an edge, and topological features at scales
         larger than this value will not be detected.
@@ -89,7 +89,7 @@ class VietorisRipsPersistence(BaseEstimator, TransformerMixin, PlotterMixin):
     Notes
     -----
     `Ripser <https://github.com/Ripser/ripser>`_ is used as a C++ backend
-    for computing Vietoris-Rips persistent homology. Python bindings were
+    for computing Vietoris–Rips persistent homology. Python bindings were
     modified for performance from the `ripser.py
     <https://github.com/scikit-tda/ripser.py>`_ package.
 
@@ -99,7 +99,7 @@ class VietorisRipsPersistence(BaseEstimator, TransformerMixin, PlotterMixin):
 
     References
     ----------
-    [1] U. Bauer, "Ripser: efficient computation of Vietoris-Rips persistence \
+    [1] U. Bauer, "Ripser: efficient computation of Vietoris–Rips persistence \
         barcodes", 2019; `arXiv:1908.02518 \
         <https://arxiv.org/abs/1908.02518>`_.
 
@@ -225,27 +225,33 @@ class VietorisRipsPersistence(BaseEstimator, TransformerMixin, PlotterMixin):
                                    self.infinity_values_, self.n_jobs)
         return Xt
 
-    def plot(self, Xt, sample=0):
-        """Plot a single persistence diagram.
+    @staticmethod
+    def plot(Xt, sample=0, homology_dimensions=None):
+        """Plot a sample from a collection of persistence diagrams, with
+        homology in multiple dimensions.
 
         Parameters
         ----------
-        Xt : ndarray of shape (n_samples, n_features, 3)
-            Array of persistence diagrams such as the result of a call to
+        Xt : ndarray of shape (n_samples, n_points, 3)
+            Collection of persistence diagrams, such as returned by
             :meth:`transform`.
 
         sample : int, optional, default: ``0``
-            Index of the sample to be plotted.
+            Index of the sample in `Xt` to be plotted.
+
+        homology_dimensions : list, tuple or None, optional, default: ``None``
+            Which homology dimensions to include in the plot. ``None`` means
+            plotting all dimensions present in ``Xt[sample]``.
 
         """
-        return plot_diagram(Xt[sample],
-                            homology_dimensions=self.homology_dimensions)
+        return plot_diagram(
+            Xt[sample], homology_dimensions=homology_dimensions)
 
 
 @adapt_fit_transform_docs
-class SparseRipsPersistence(BaseEstimator, TransformerMixin):
+class SparseRipsPersistence(BaseEstimator, TransformerMixin, PlotterMixin):
     """:ref:`Persistence diagrams <persistence diagram>` resulting from
-    :ref:`Sparse Vietoris-Rips filtrations <vietoris-rips complex and \
+    :ref:`Sparse Vietoris–Rips filtrations <vietoris-rips complex and \
     vietoris-rips homology>`.
 
     Given a :ref:`point cloud <finite metric spaces and point clouds>` in
@@ -283,12 +289,12 @@ class SparseRipsPersistence(BaseEstimator, TransformerMixin):
         :math:`p` equals `coeff`.
 
     epsilon : float between 0. and 1., optional, default: ``0.1``
-        Parameter controlling the approximation to the exact Vietoris-Rips
+        Parameter controlling the approximation to the exact Vietoris–Rips
         filtration. If set to `0.`, :class:`SparseRipsPersistence` leads to
         the same results as :class:`VietorisRipsPersistence` but is slower.
 
     max_edge_length : float, optional, default: ``numpy.inf``
-        Upper bound on the maximum value of the Vietoris-Rips filtration
+        Upper bound on the maximum value of the Vietoris–Rips filtration
         parameter. Points whose distance is greater than this value will
         never be connected by an edge, and topological features at scales
         larger than this value will not be detected.
@@ -317,7 +323,7 @@ class SparseRipsPersistence(BaseEstimator, TransformerMixin):
     Notes
     -----
     `GUDHI <https://github.com/GUDHI/gudhi-devel>`_ is used as a C++ backend
-    for computing sparse Vietoris-Rips persistent homology. Python bindings
+    for computing sparse Vietoris–Rips persistent homology. Python bindings
     were modified for performance.
 
     Persistence diagrams produced by this class must be interpreted with
@@ -464,9 +470,31 @@ class SparseRipsPersistence(BaseEstimator, TransformerMixin):
                                    self.infinity_values_, self.n_jobs)
         return Xt
 
+    @staticmethod
+    def plot(Xt, sample=0, homology_dimensions=None):
+        """Plot a sample from a collection of persistence diagrams, with
+        homology in multiple dimensions.
+
+        Parameters
+        ----------
+        Xt : ndarray of shape (n_samples, n_points, 3)
+            Collection of persistence diagrams, such as returned by
+            :meth:`transform`.
+
+        sample : int, optional, default: ``0``
+            Index of the sample in `Xt` to be plotted.
+
+        homology_dimensions : list, tuple or None, optional, default: ``None``
+            Which homology dimensions to include in the plot. ``None`` means
+            plotting all dimensions present in ``Xt[sample]``.
+
+        """
+        return plot_diagram(
+            Xt[sample], homology_dimensions=homology_dimensions)
+
 
 @adapt_fit_transform_docs
-class EuclideanCechPersistence(BaseEstimator, TransformerMixin):
+class EuclideanCechPersistence(BaseEstimator, TransformerMixin, PlotterMixin):
     """:ref:`Persistence diagrams <persistence diagram>` resulting from
     `Cech filtrations <>`_.
 
@@ -488,7 +516,7 @@ class EuclideanCechPersistence(BaseEstimator, TransformerMixin):
         :math:`p` equals `coeff`.
 
     max_edge_length : float, optional, default: ``numpy.inf``
-        Upper bound on the maximum value of the Vietoris-Rips filtration
+        Upper bound on the maximum value of the Vietoris–Rips filtration
         parameter. Points whose distance is greater than this value will
         never be connected by an edge, and topological features at scales
         larger than this value will not be detected.
@@ -649,3 +677,25 @@ class EuclideanCechPersistence(BaseEstimator, TransformerMixin):
         Xt = _postprocess_diagrams(Xt, self._homology_dimensions,
                                    self.infinity_values_, self.n_jobs)
         return Xt
+
+    @staticmethod
+    def plot(Xt, sample=0, homology_dimensions=None):
+        """Plot a sample from a collection of persistence diagrams, with
+        homology in multiple dimensions.
+
+        Parameters
+        ----------
+        Xt : ndarray of shape (n_samples, n_points, 3)
+            Collection of persistence diagrams, such as returned by
+            :meth:`transform`.
+
+        sample : int, optional, default: ``0``
+            Index of the sample in `Xt` to be plotted.
+
+        homology_dimensions : list, tuple or None, optional, default: ``None``
+            Which homology dimensions to include in the plot. ``None`` means
+            plotting all dimensions present in ``Xt[sample]``.
+
+        """
+        return plot_diagram(
+            Xt[sample], homology_dimensions=homology_dimensions)
