+++ gtda/images/filtrations.py
@@ -1,4 +1,4 @@
-"""Binary image filtration."""
+"""Filtrations of 2D/3D binary images."""
 # License: GNU AGPLv3
 
 from numbers import Real
@@ -13,17 +13,19 @@ from sklearn.utils import gen_even_slices
 from sklearn.utils.validation import check_is_fitted, check_array
 
 from ._utils import _dilate, _erode
+from ..base import PlotterMixin
+from ..plotting import plot_heatmap
 from ..utils._docs import adapt_fit_transform_docs
 from ..utils.intervals import Interval
 from ..utils.validation import validate_params
 
 
 @adapt_fit_transform_docs
-class HeightFiltration(BaseEstimator, TransformerMixin):
+class HeightFiltration(BaseEstimator, TransformerMixin, PlotterMixin):
     """Filtrations of 2D/3D binary images based on distances to lines/planes.
 
     The height filtration assigns to each activated pixel of a binary image a
-    grayscale value equal to the distance between the pixel and the hyperplane
+    greyscale value equal to the distance between the pixel and the hyperplane
     defined by a direction vector and the first seen edge of the image
     following that direction. Deactivated pixels are assigned the value of the
     maximum distance between any pixel of the image and the hyperplane, plus
@@ -53,7 +55,7 @@ class HeightFiltration(BaseEstimator, TransformerMixin):
         Effective direction of the height filtration. Set in :meth:`fit`.
 
     mesh_ : ndarray of shape ( n_pixels_x, n_pixels_y [, n_pixels_z])
-        Grayscale image corresponding to the height filtration of a binary
+        greyscale image corresponding to the height filtration of a binary
         image where each pixel is activated. Set in :meth:`fit`.
 
     max_value_: float
@@ -138,10 +140,10 @@ class HeightFiltration(BaseEstimator, TransformerMixin):
 
     def transform(self, X, y=None):
         """For each binary image in the collection `X`, calculate a
-        corresponding grayscale image based on the distance of its pixels to
+        corresponding greyscale image based on the distance of its pixels to
         the hyperplane defined by the `direction` vector and the first seen
         edge of the images following that `direction`. Return the collection
-        of grayscale images.
+        of greyscale images.
 
         Parameters
         ----------
@@ -158,7 +160,7 @@ class HeightFiltration(BaseEstimator, TransformerMixin):
         Xt : ndarray of shape (n_samples, n_pixels_x,
             n_pixels_y [, n_pixels_z])
             Transformed collection of images. Each entry along axis 0 is a
-            2D or 3D grayscale image.
+            2D or 3D greyscale image.
 
         """
         check_is_fitted(self)
@@ -172,13 +174,38 @@ class HeightFiltration(BaseEstimator, TransformerMixin):
 
         return Xt
 
+    @staticmethod
+    def plot(Xt, sample=0, colorscale='greys', origin='upper'):
+        """Plot a sample from a collection of 2D greyscale images.
+
+        Parameters
+        ----------
+        Xt : ndarray of shape (n_samples, n_pixels_x, n_pixels_y)
+            Collection of 2D greyscale images, such as returned by
+            :meth:`transform`.
+
+        sample : int, optional, default: ``0``
+            Index of the sample in `Xt` to be plotted.
+
+        colorscale : str, optional, default: ``'greys'``
+            Color scale to be used in the heat map. Can be anything allowed by
+            :class:`plotly.graph_objects.Heatmap`.
+
+        origin : ``'upper'`` | ``'lower'``, optional, default: ``'upper'``
+            Position of the [0, 0] pixel of `data`, in the upper left or lower
+            left corner. The convention ``'upper'`` is typically used for
+            matrices and images.
+
+        """
+        return plot_heatmap(Xt[sample], colorscale=colorscale, origin=origin)
+
 
 @adapt_fit_transform_docs
-class RadialFiltration(BaseEstimator, TransformerMixin):
+class RadialFiltration(BaseEstimator, TransformerMixin, PlotterMixin):
     """Filtrations of 2D/3D binary images based on distances to a reference
     pixel.
 
-    The radial filtration assigns to each pixel of a binary image a grayscale
+    The radial filtration assigns to each pixel of a binary image a greyscale
     value computed as follows in terms of a reference pixel, called the
     "center", and of a "radius": if the binary pixel is active and lies
     within a ball defined by this center and this radius, then the assigned
@@ -235,7 +262,7 @@ class RadialFiltration(BaseEstimator, TransformerMixin):
         the empty dictionary.
 
     mesh_ : ndarray of shape ( n_pixels_x, n_pixels_y [, n_pixels_z])
-        Grayscale image corresponding to the radial filtration of a binary
+        greyscale image corresponding to the radial filtration of a binary
         image where each pixel is activated. Set in :meth:`fit`.
 
     max_value_: float
@@ -335,8 +362,8 @@ class RadialFiltration(BaseEstimator, TransformerMixin):
 
     def transform(self, X, y=None):
         """For each binary image in the collection `X`, calculate a
-        corresponding grayscale image based on the distance of its pixels to
-        the center. Return the collection of grayscale images.
+        corresponding greyscale image based on the distance of its pixels to
+        the center. Return the collection of greyscale images.
 
         Parameters
         ----------
@@ -353,7 +380,7 @@ class RadialFiltration(BaseEstimator, TransformerMixin):
         Xt : ndarray of shape (n_samples, n_pixels_x,
             n_pixels_y [, n_pixels_z])
             Transformed collection of images. Each entry along axis 0 is a
-            2D or 3D grayscale image.
+            2D or 3D greyscale image.
 
         """
         check_is_fitted(self)
@@ -367,9 +394,34 @@ class RadialFiltration(BaseEstimator, TransformerMixin):
 
         return Xt
 
+    @staticmethod
+    def plot(Xt, sample=0, colorscale='greys', origin='upper'):
+        """Plot a sample from a collection of 2D greyscale images.
+
+        Parameters
+        ----------
+        Xt : ndarray of shape (n_samples, n_pixels_x, n_pixels_y)
+            Collection of 2D greyscale images, such as returned by
+            :meth:`transform`.
+
+        sample : int, optional, default: ``0``
+            Index of the sample in `Xt` to be plotted.
+
+        colorscale : str, optional, default: ``'greys'``
+            Color scale to be used in the heat map. Can be anything allowed by
+            :class:`plotly.graph_objects.Heatmap`.
+
+        origin : ``'upper'`` | ``'lower'``, optional, default: ``'upper'``
+            Position of the [0, 0] pixel of `data`, in the upper left or lower
+            left corner. The convention ``'upper'`` is typically used for
+            matrices and images.
+
+        """
+        return plot_heatmap(Xt[sample], colorscale=colorscale, origin=origin)
+
 
 @adapt_fit_transform_docs
-class DilationFiltration(BaseEstimator, TransformerMixin):
+class DilationFiltration(BaseEstimator, TransformerMixin, PlotterMixin):
     """Filtrations of 2D/3D binary images based on the dilation of activated
     regions.
 
@@ -377,12 +429,12 @@ class DilationFiltration(BaseEstimator, TransformerMixin):
     image processing and relies on the `scipy.ndimage \
     <https://docs.scipy.org/doc/scipy/reference/ndimage.html>`_ module.
 
-    This filtration assigns to each pixel in an image a grayscale value
+    This filtration assigns to each pixel in an image a greyscale value
     calculated as follows. If the minimum Manhattan distance between the
     pixel and any activated pixel in the image is less than or equal to
     the parameter `n_iterations`, the assigned value is this distance –
     in particular, activated pixels are assigned a value of 0.
-    Otherwise, the assigned grayscale value is the sum of the lengths
+    Otherwise, the assigned greyscale value is the sum of the lengths
     along all axes of the image – equivalently, it is the maximum
     Manhattan distance between any two pixels in the image. The name of
     this filtration comes from the fact that these values can be computed
@@ -475,9 +527,9 @@ class DilationFiltration(BaseEstimator, TransformerMixin):
 
     def transform(self, X, y=None):
         """For each binary image in the collection `X`, calculate a
-        corresponding grayscale image based on the distance of its pixels to
+        corresponding greyscale image based on the distance of its pixels to
         their closest activated neighboring pixel. Return the collection
-        of grayscale images.
+        of greyscale images.
 
         Parameters
         ----------
@@ -494,7 +546,7 @@ class DilationFiltration(BaseEstimator, TransformerMixin):
         Xt : ndarray of shape (n_samples, n_pixels_x,
             n_pixels_y [, n_pixels_z])
             Transformed collection of images. Each entry along axis 0 is a
-            2D or 3D grayscale image.
+            2D or 3D greyscale image.
 
         """
         check_is_fitted(self)
@@ -508,9 +560,34 @@ class DilationFiltration(BaseEstimator, TransformerMixin):
 
         return Xt
 
+    @staticmethod
+    def plot(Xt, sample=0, colorscale='greys', origin='upper'):
+        """Plot a sample from a collection of 2D greyscale images.
+
+        Parameters
+        ----------
+        Xt : ndarray of shape (n_samples, n_pixels_x, n_pixels_y)
+            Collection of 2D greyscale images, such as returned by
+            :meth:`transform`.
+
+        sample : int, optional, default: ``0``
+            Index of the sample in `Xt` to be plotted.
+
+        colorscale : str, optional, default: ``'greys'``
+            Color scale to be used in the heat map. Can be anything allowed by
+            :class:`plotly.graph_objects.Heatmap`.
+
+        origin : ``'upper'`` | ``'lower'``, optional, default: ``'upper'``
+            Position of the [0, 0] pixel of `data`, in the upper left or lower
+            left corner. The convention ``'upper'`` is typically used for
+            matrices and images.
+
+        """
+        return plot_heatmap(Xt[sample], colorscale=colorscale, origin=origin)
+
 
 @adapt_fit_transform_docs
-class ErosionFiltration(BaseEstimator, TransformerMixin):
+class ErosionFiltration(BaseEstimator, TransformerMixin, PlotterMixin):
     """Filtrations of 2D/3D binary images based on the erosion of activated
     regions.
 
@@ -518,12 +595,12 @@ class ErosionFiltration(BaseEstimator, TransformerMixin):
     image processing and relies on the `scipy.ndimage \
     <https://docs.scipy.org/doc/scipy/reference/ndimage.html>`_ module.
 
-    This filtration assigns to each pixel in an image a grayscale value
+    This filtration assigns to each pixel in an image a greyscale value
     calculated as follows. If the minimum Manhattan distance between the
     pixel and any deactivated pixel in the image is less than or equal to
     the parameter `n_iterations`, the assigned value is this distance –
     in particular, deactivated pixels are assigned a value of 0.
-    Otherwise, the assigned grayscale value is the sum of the lengths
+    Otherwise, the assigned greyscale value is the sum of the lengths
     along all axes of the image – equivalently, it is the maximum
     Manhattan distance between any two pixels in the image. The name of
     this filtration comes from the fact that these values can be computed
@@ -616,9 +693,9 @@ class ErosionFiltration(BaseEstimator, TransformerMixin):
 
     def transform(self, X, y=None):
         """For each binary image in the collection `X`, calculate a
-        corresponding grayscale image based on the distance of its pixels to
+        corresponding greyscale image based on the distance of its pixels to
         their closest activated neighboring pixel. Return the collection
-        of grayscale images.
+        of greyscale images.
 
         Parameters
         ----------
@@ -635,7 +712,7 @@ class ErosionFiltration(BaseEstimator, TransformerMixin):
         Xt : ndarray of shape (n_samples, n_pixels_x,
             n_pixels_y [, n_pixels_z])
             Transformed collection of images. Each entry along axis 0 is a
-            2D or 3D grayscale image.
+            2D or 3D greyscale image.
 
         """
         check_is_fitted(self)
@@ -649,23 +726,48 @@ class ErosionFiltration(BaseEstimator, TransformerMixin):
 
         return Xt
 
+    @staticmethod
+    def plot(Xt, sample=0, colorscale='greys', origin='upper'):
+        """Plot a sample from a collection of 2D greyscale images.
+
+        Parameters
+        ----------
+        Xt : ndarray of shape (n_samples, n_pixels_x, n_pixels_y)
+            Collection of 2D greyscale images, such as returned by
+            :meth:`transform`.
+
+        sample : int, optional, default: ``0``
+            Index of the sample in `Xt` to be plotted.
+
+        colorscale : str, optional, default: ``'greys'``
+            Color scale to be used in the heat map. Can be anything allowed by
+            :class:`plotly.graph_objects.Heatmap`.
+
+        origin : ``'upper'`` | ``'lower'``, optional, default: ``'upper'``
+            Position of the [0, 0] pixel of `data`, in the upper left or lower
+            left corner. The convention ``'upper'`` is typically used for
+            matrices and images.
+
+        """
+        return plot_heatmap(Xt[sample], colorscale=colorscale, origin=origin)
+
 
 @adapt_fit_transform_docs
-class SignedDistanceFiltration(BaseEstimator, TransformerMixin):
+class SignedDistanceFiltration(BaseEstimator, TransformerMixin, PlotterMixin):
     """Filtrations of 2D/3D binary images based on the dilation and the erosion
     of activated regions.
 
-    This filtration assigns to each pixel in an image a grayscale value
+    This filtration assigns to each pixel in an image a greyscale value
     calculated as follows. For activated pixels, if the minimum Manhattan
     distance between the pixel and any deactivated pixel in the image is less
     than or equal to the parameter `n_iterations`, the assigned value is
-    this distance minus 1. Otherwise, the assigned grayscale value is the sum
+    this distance minus 1. Otherwise, the assigned greyscale value is the sum
     of the lengths along all axes of the image – equivalently, it is the
     maximum Manhattan distance between any two pixels in the image, minus 1.
     For deactivated pixels, if the minimum Manhattan distance between the pixel
     and any activated pixel in the image is less than or equal to the parameter
     `n_iterations`, the assigned value is the opposite of this distance.
-    Otherwise, the assigned grayscale value is the opposite of the maximum
+    Otherwise, the assigned greyscale value is the opposite of the maximum
     Manhattan distance between any two pixels in the image.
 
     The name of this filtration comes from the fact that it is a a negatively
@@ -767,9 +869,9 @@ class SignedDistanceFiltration(BaseEstimator, TransformerMixin):
 
     def transform(self, X, y=None):
         """For each binary image in the collection `X`, calculate a
-        corresponding grayscale image based on the distance of its pixels to
+        corresponding greyscale image based on the distance of its pixels to
         their closest activated neighboring pixel. Return the collection
-        of grayscale images.
+        of greyscale images.
 
         Parameters
         ----------
@@ -786,7 +888,7 @@ class SignedDistanceFiltration(BaseEstimator, TransformerMixin):
         Xt : ndarray of shape (n_samples, n_pixels_x,
             n_pixels_y [, n_pixels_z])
             Transformed collection of images. Each entry along axis 0 is a
-            2D or 3D grayscale image.
+            2D or 3D greyscale image.
 
         """
         check_is_fitted(self)
@@ -799,3 +901,28 @@ class SignedDistanceFiltration(BaseEstimator, TransformerMixin):
         Xt = np.concatenate(Xt)
 
         return Xt
+
+    @staticmethod
+    def plot(Xt, sample=0, colorscale='greys', origin='upper'):
+        """Plot a sample from a collection of 2D greyscale images.
+
+        Parameters
+        ----------
+        Xt : ndarray of shape (n_samples, n_pixels_x, n_pixels_y)
+            Collection of 2D greyscale images, such as returned by
+            :meth:`transform`.
+
+        sample : int, optional, default: ``0``
+            Index of the sample in `Xt` to be plotted.
+
+        colorscale : str, optional, default: ``'greys'``
+            Color scale to be used in the heat map. Can be anything allowed by
+            :class:`plotly.graph_objects.Heatmap`.
+
+        origin : ``'upper'`` | ``'lower'``, optional, default: ``'upper'``
+            Position of the [0, 0] pixel of `data`, in the upper left or lower
+            left corner. The convention ``'upper'`` is typically used for
+            matrices and images.
+
+        """
+        return plot_heatmap(Xt[sample], colorscale=colorscale, origin=origin)
