+++ gtda/images/preprocessing.py
@@ -10,23 +10,22 @@ from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.utils import gen_even_slices
 from sklearn.utils.validation import check_is_fitted, check_array
 
+from ..base import PlotterMixin
+from ..plotting import plot_point_cloud, plot_heatmap
 from ..utils._docs import adapt_fit_transform_docs
 from ..utils.intervals import Interval
 from ..utils.validation import validate_params
 
 
 @adapt_fit_transform_docs
-class Binarizer(BaseEstimator, TransformerMixin):
-    """Binarize all 2D/3D grayscale images in a collection.
+class Binarizer(BaseEstimator, TransformerMixin, PlotterMixin):
+    """Binarize all 2D/3D greyscale images in a collection.
 
     Parameters
     ----------
     threshold : float, default: 0.5
-        Percentage of the maximum pixel value `max_value_` from which to
+        Fraction of the maximum pixel value `max_value_` from which to
         binarize.
-
-    normalize: bool, optional, default: ``False``
-        If ``True``, divide every pixel value by `max_value_`.
 
     n_jobs : int or None, optional, default: ``None``
         The number of jobs to use for the computation. ``None`` means 1 unless
@@ -49,26 +48,21 @@ class Binarizer(BaseEstimator, TransformerMixin):
     """
 
     _hyperparameters = {
-        'threshold': {'type': Real, 'in': Interval(0, 1, closed='right')},
-        'normalize': {'type': bool}
+        'threshold': {'type': Real, 'in': Interval(0, 1, closed='right')}
     }
 
-    def __init__(self, threshold=0.5, normalize=False, n_jobs=None):
+    def __init__(self, threshold=0.5, n_jobs=None):
         self.threshold = threshold
-        self.normalize = normalize
         self.n_jobs = n_jobs
 
     def _binarize(self, X):
         Xbin = X / self.max_value_ > self.threshold
 
-        if self.normalize:
-            Xbin = Xbin * self.max_value_
-
         return Xbin
 
     def fit(self, X, y=None):
         """Calculate :attr:`n_dimensions_` and :attr:`max_value_` from the
-        collection of grayscale images. Then, return the estimator.
+        collection of greyscale images. Then, return the estimator.
 
         This method is here to implement the usual scikit-learn API and hence
         work in pipelines.
@@ -78,7 +72,7 @@ class Binarizer(BaseEstimator, TransformerMixin):
         X : ndarray of shape (n_samples, n_pixels_x, n_pixels_y \
             [, n_pixels_z])
             Input data. Each entry along axis 0 is interpreted as a 2D or 3D
-            grayscale image.
+            greyscale image.
 
         y : None
             There is no need of a target in a transformer, yet the pipeline API
@@ -102,7 +96,7 @@ class Binarizer(BaseEstimator, TransformerMixin):
         return self
 
     def transform(self, X, y=None):
-        """For each grayscale image in the collection `X`, calculate a
+        """For each greyscale image in the collection `X`, calculate a
         corresponding binary image by applying the `threshold`. Return the
         collection of binary images.
 
@@ -110,7 +104,7 @@ class Binarizer(BaseEstimator, TransformerMixin):
         ----------
         X : ndarray of shape (n_samples, n_pixels_x, n_pixels_y [, n_pixels_z])
             Input data. Each entry along axis 0 is interpreted as a 2D or 3D
-            grayscale image.
+            greyscale image.
 
         y : None
             There is no need of a target in a transformer, yet the pipeline API
@@ -138,9 +132,35 @@ class Binarizer(BaseEstimator, TransformerMixin):
 
         return Xt
 
+    @staticmethod
+    def plot(Xt, sample=0, colorscale='greys', origin='upper'):
+        """Plot a sample from a collection of 2D binary images.
+
+        Parameters
+        ----------
+        Xt : ndarray of shape (n_samples, n_pixels_x, n_pixels_y)
+            Collection of 2D binary images, such as returned by
+            :meth:`transform`.
+
+        sample : int, optional, default: ``0``
+            Index of the sample in `Xt` to be plotted.
+
+        colorscale : str, optional, default: ``'greys'``
+            Color scale to be used in the heat map. Can be anything allowed by
+            :class:`plotly.graph_objects.Heatmap`.
+
+        origin : ``'upper'`` | ``'lower'``, optional, default: ``'upper'``
+            Position of the [0, 0] pixel of `data`, in the upper left or lower
+            left corner. The convention ``'upper'`` is typically used for
+            matrices and images.
+
+        """
+        return plot_heatmap(
+            Xt[sample] * 1, colorscale=colorscale, origin=origin)
+
 
 @adapt_fit_transform_docs
-class Inverter(BaseEstimator, TransformerMixin):
+class Inverter(BaseEstimator, TransformerMixin, PlotterMixin):
     """Invert all 2D/3D binary images in a collection.
 
     Parameters
@@ -214,9 +234,35 @@ class Inverter(BaseEstimator, TransformerMixin):
 
         return Xt
 
+    @staticmethod
+    def plot(Xt, sample=0, colorscale='greys', origin='upper'):
+        """Plot a sample from a collection of 2D binary images.
+
+        Parameters
+        ----------
+        Xt : ndarray of shape (n_samples, n_pixels_x, n_pixels_y)
+            Collection of 2D binary images, such as returned by
+            :meth:`transform`.
+
+        sample : int, optional, default: ``0``
+            Index of the sample in `Xt` to be plotted.
+
+        colorscale : str, optional, default: ``'greys'``
+            Color scale to be used in the heat map. Can be anything allowed by
+            :class:`plotly.graph_objects.Heatmap`.
+
+        origin : ``'upper'`` | ``'lower'``, optional, default: ``'upper'``
+            Position of the [0, 0] pixel of `data`, in the upper left or lower
+            left corner. The convention ``'upper'`` is typically used for
+            matrices and images.
+
+        """
+        return plot_heatmap(
+            Xt[sample] * 1, colorscale=colorscale, origin=origin)
+
 
 @adapt_fit_transform_docs
-class Padder(BaseEstimator, TransformerMixin):
+class Padder(BaseEstimator, TransformerMixin, PlotterMixin):
     """Pad all 2D/3D binary images in a collection.
 
     Parameters
@@ -334,9 +380,35 @@ class Padder(BaseEstimator, TransformerMixin):
 
         return Xt
 
+    @staticmethod
+    def plot(Xt, sample=0, colorscale='greys', origin='upper'):
+        """Plot a sample from a collection of 2D binary images.
+
+        Parameters
+        ----------
+        Xt : ndarray of shape (n_samples, n_pixels_x, n_pixels_y)
+            Collection of 2D binary images, such as returned by
+            :meth:`transform`.
+
+        sample : int, optional, default: ``0``
+            Index of the sample in `Xt` to be plotted.
+
+        colorscale : str, optional, default: ``'greys'``
+            Color scale to be used in the heat map. Can be anything allowed by
+            :class:`plotly.graph_objects.Heatmap`.
+
+        origin : ``'upper'`` | ``'lower'``, optional, default: ``'upper'``
+            Position of the [0, 0] pixel of `data`, in the upper left or lower
+            left corner. The convention ``'upper'`` is typically used for
+            matrices and images.
+
+        """
+        return plot_heatmap(
+            Xt[sample] * 1, colorscale=colorscale, origin=origin)
+
 
 @adapt_fit_transform_docs
-class ImageToPointCloud(BaseEstimator, TransformerMixin):
+class ImageToPointCloud(BaseEstimator, TransformerMixin, PlotterMixin):
     """Represent active pixels in 2D/3D binary images as points in 2D/3D space.
 
     The coordinates of each point is calculated as follows. For each activated
@@ -444,3 +516,20 @@ class ImageToPointCloud(BaseEstimator, TransformerMixin):
                                      effective_n_jobs(self.n_jobs)))
         Xt = np.concatenate(Xt)
         return Xt
+
+    @staticmethod
+    def plot(Xt, sample=0):
+        """Plot a sample from a collection of point clouds. If the point cloud
+        is in more than three dimensions, only the first three are plotted.
+
+        Parameters
+        ----------
+        Xt : ndarray, shape (n_samples, n_points, n_dimensions)
+            Collection of point clouds in ``n_dimension``-dimensional space,
+            such as returned by :meth:`transform`.
+
+        sample : int, optional, default: ``0``
+            Index of the sample in `Xt` to be plotted.
+
+        """
+        return plot_point_cloud(Xt[sample])
