+++ gtda/utils/__init__.py
@@ -1,12 +1,12 @@
 """The module :mod:`gtda.utils` implements hyperparameter and input
 validation functions."""
 
-from .validation import check_diagram, check_graph
+from .validation import check_diagrams, check_graph
 from .validation import validate_params
 
 
 __all__ = [
-    'check_diagram',
+    'check_diagrams',
     'check_graph',
     'validate_params'
 ]
